+++ docs/conf.py
@@ -3,7 +3,7 @@ from   permutation import __version__
 
 project   = 'permutation'
 author    = 'John T. Wodder II'
-copyright = '2017 John T. Wodder II'
+copyright = '2017-2019 John T. Wodder II'
 
 extensions = [
     'sphinx.ext.autodoc',